test(tools): share the skill manager stub and trim comments
